Fix admin script user name (#3735)
After #3519, we are getting the user_name from the "fed_admin.json":
https://github.com/NVIDIA/NVFlare/blob/main/nvflare/fuel/hci/tools/admin.py#L91

But our provision is not updated to include this "user_name".
As a result, no matter what users specify in "project.yml"
The admin user name is always the default "admin"

This PR fix it.

### Description

- Add the user_name section in static file builder and template

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

Co-authored-by: Chester Chen <512707+chesterxgchen@users.noreply.github.com>