Fix job process start commands (#3733)
Fixes # .

### Description

The commands for starting SJ and CJ need to adding extra options to the
args. Currently it directly works on the globally saved args. Every time
a new CJ/SJ is started, the args are added extra options, which cause
the command to have more and more such options.

This PR changes to use a copy of the global args for the operation.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

Co-authored-by: Yuan-Ting Hsieh (謝沅廷) <yuantingh@nvidia.com>