Reset task_data in client after executor is run. (#3763)
This fixes https://nvbugspro.nvidia.com/bug/5570625

### Description

The client relies on task_data in executor. This change clears the
task_data from FLContext after the executor is run.

The default file streaming size is changed to 0 so it has the same
behavior as 2.6.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

---------

Co-authored-by: Chester Chen <512707+chesterxgchen@users.noreply.github.com>