Address race issue for quantization filter (#3865)
Fixes # .

### Description

this is a new issue that has not been reported - previously we addressed
the "double filter" issue that an already filtered result get filtered
again later by another request - this is assuming requests are
relatively sequential - but under multi-thread, the request can happen
at the same time causing racing issue when it takes time for the flag to
be set, and our previous solution will not be able to address that.

New solution involves lock and proper logic to wait for the correct
filtered result to be produced

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.