Consolidate BaseFedJob and fedavg.py (#3879)
Consolidates BaseFedJob and fedavg.py.

### Description

This first step creates a new unified BaseFedJob and fedavg.py based off
of UnifiedFedAvgRecipe and wrappers to make the existing code work as
is. We can continue to modify the different places it is used to
directly use the unified implementations.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

---------

Co-authored-by: Chester Chen <512707+chesterxgchen@users.noreply.github.com>